# swagger.yml
swagger: '2.0'
info:
  description: This document describe the API for transactions App. A financial transactions tracking.
  version: 1.0.0
  title: Swagger Transactions App
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: renatto.martins@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: transações.io
basePath: /v1
...
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/accounts/{accountId}/transactions':
    get:
      summary: Get a list of transactions
      tags:
        - transactions
      operationId: getTransactions
      parameters:
        - name: accountId
          in: path
          description: ID of account to which the transactions belong
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Ok
          headers:
            X-Total-Count:
              type: integer
              format: int32
              description: Number of transactions in response
          schema:
            $ref: '#/definitions/TransactionList'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/400CodeResponsePayload'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/401CodeResponsePayload'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/403CodeResponsePayload'
        '404':
          description: Account not found
          schema:
            $ref: '#/definitions/404CodeResponsePayload'
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/500CodeResponsePayload'
    post:
      tags:
        - transactions
      summary: Add a new transaction
      description: ''
      operationId: addTransaction
      parameters:
        - name: accountId
          in: path
          description: ID of account to which the transaction belongs
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Transaction object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/TransactionRequestPayload'
      responses:
        '201':
          description: Created
          headers:
            Location:
              type: string
              description: The location of the new Transaction
          schema:
            $ref: '#/definitions/Transaction'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/400CodeResponsePayload'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/401CodeResponsePayload'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/403CodeResponsePayload'
        '404':
          description: Account not found
          schema:
            $ref: '#/definitions/404CodeResponsePayload'
        '422':
          description: 'Unprocessable Entity'
          schema:
            $ref: '#/definitions/422CodeResponsePayload'
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/500CodeResponsePayload'
  '/accounts/{accountId}/transactions/{transactionId}':
    get:
      tags:
        - transactions
      summary: Get a transaction
      operationId: getTransaction
      parameters:
        - name: accountId
          in: path
          description: ID of account to which the transaction belongs
          required: true
          type: integer
          format: int64
        - name: transactionId
          in: path
          description: ID of transaction to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Transaction'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/400CodeResponsePayload'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/401CodeResponsePayload'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/403CodeResponsePayload'
        '404':
          description: Account ou Transaction not found
          schema:
            $ref: '#/definitions/404CodeResponsePayload'
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/500CodeResponsePayload'
    put:
      tags:
        - transactions
      summary: Update an existing transaction
      operationId: updateTransaction
      parameters:
        - name: accountId
          in: path
          description: ID of account to which the transaction belongs
          required: true
          type: integer
          format: int64
        - name: transactionId
          in: path
          description: ID of transaction to be updated
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Transaction object that needs to be uptaded
          required: true
          schema:
            $ref: '#/definitions/TransactionRequestPayload'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Transaction'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/400CodeResponsePayload'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/401CodeResponsePayload'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/403CodeResponsePayload'
        '404':
          description: Transaction not found
          schema:
            $ref: '#/definitions/404CodeResponsePayload'
        '422':
          description: 'Unprocessable Entity'
          schema:
            $ref: '#/definitions/422CodeResponsePayload'
        '500':
          description: Internal server error
          schema:
            $ref: '#/definitions/500CodeResponsePayload'
    patch:
      tags:
        - transactions
      summary: Update parcially an existing transaction
      operationId: patchTransaction
      parameters:
        - name: accountId
          in: path
          description: ID of account to which the transaction belongs
          required: true
          type: integer
          format: int64
        - name: transactionId
          in: path
          description: ID of transaction to return
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Transaction object that needs to be uptaded
          required: true
          schema:
            $ref: '#/definitions/Transaction'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Transaction'
        '400':
          description: Invalid ID supplied
        '404':
          description: Transaction not found
        '405':
          description: Validation exception
    delete:
      tags:
        - transactions
      summary: Delete an existing transaction
      operationId: deleteTransaction
      parameters:
        - name: accountId
          in: path
          description: ID of account to which the transaction belongs
          required: true
          type: integer
          format: int64
        - name: transactionId
          in: path
          description: ID of transaction to return
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: No content
        '400':
          description: Invalid ID supplied
        '404':
          description: Transaction not found
definitions:
...
  500CodeResponsePayload:
    type: object
    properties:
      code:
        type: integer
        example: 500
      message:
        type: string
        example: 'Internal Server Error'
  400CodeResponsePayload:
    type: object
    properties:
      code:
        type: integer
        example: 400
      message:
        type: string
        example: 'Bad request'
  401CodeResponsePayload:
    type: object
    properties:
      code:
        type: integer
        example: 401
      message:
        type: string
        example: 'Unauthorized'
  403CodeResponsePayload:
    type: object
    properties:
      code:
        type: integer
        example: 403
      message:
        type: string
        example: 'Forbidden'
  404CodeResponsePayload:
    type: object
    properties:
      code:
        type: integer
        example: 404
      message:
        type: string
        example: 'Not found'
...
  422CodeResponsePayload:
    type: object
    properties:
      code:
        type: integer
        example: 422
      message:
        type: string
        example: 'Unprocessable Entity'
      details:
        type: array
        items:
          type: object
          properties:
            type:
              type: string
              example: 'field'
            msg:
              type: string
              example: 'invalid field'
            path:
              type: string
              example: '<field_name>'
            location:
              type: string
              example: 'body'
...
  TransactionRequestPayload:
    type: object
    required:
      - description
      - amount
      - date
      - isIncome
    properties:
      description:
        type: string
        example: Conta de água
      amount:
        type: number
        format: double
        example: 128.32
      date:
        type: string
        example: "2023-04-04T08:14:34.606Z"
      notes:
        type: string
        example: "Conta referente ao mês jul/2023"
      isIncome:
        type: boolean
  TransactionList:
    type: array
    items:
      $ref: '#/definitions/Transaction'
  Transaction:
    type: object
    required:
      - description
      - amount
      - date
      - isIncome
      - isTransfer
      - accountId
    properties:
      id:
        type: integer
        format: int64
        example: 1254
      description:
        type: string
        example: Conta de água
      amount:
        type: number
        format: double
        example: 128.32
      date:
        type: string
        example: "2023-04-04T08:14:34.606Z"
      notes:
        type: string
        example: "Conta referente ao mês jul/2023"
      isIncome:
        type: boolean
      createdAt:
        type: string
        example: '2023-04-04T08:14:34.606Z'
      updatedAt:
        type: string
        example: '2023-04-04T08:14:34.606Z'
      AccountId:
        type: integer
        format: int64
        example: 4
...